refactor: major refactor part 2
