+++ src/components/Garage/Garage.module.scss
@@ -1,0 +1,33 @@
+.car__img {
+  width: 80px;
+  height: 40px;
+  mask-size: cover;
+  mask-position: center;
+  -webkit-mask-image: url(../../img/car.svg);
+  mask-image: url(../../img/car.svg);
+}
+
+@keyframes animateCar {
+  from {
+    transform: translateX(0);
+  }
+
+  to {
+    transform: translateX(calc(100vw - 120px));
+  }
+}
+
+.car__img_drive {
+  animation: animateCar linear;
+  animation-fill-mode: forwards;
+}
+
+.car__img_broken {
+  animation: animateCar linear;
+  animation-play-state: paused;
+}
+
+.hidden {
+  height: 0;
+  overflow: hidden;
+}
